Add modules folder, add reassignIndex functionality
